Exclude attributes set to None
Previously we used exclude_unset.
That doesn't work because the identifier and the schemaKey
attributes are set by default value and are thus
technically never explicitly set, and are
therefore excluded.

We now have a test to check that they are in the output.
And we also don't exclude them anymore